reverted class org structure
