+++ lib/ruby-hl7.rb
@@ -54,7 +54,7 @@ require 'segment_list_storage'
 require 'segment_generator'
 require 'segment_fields'
 require 'segment'
-# require 'segment_default'
+require 'segment_default'
 
 require 'core_ext/date_time'
 require 'core_ext/string'
