+++ lib/segment_default.rb
@@ -6,13 +6,13 @@
 #  seg.e1 = "SOMETHING ELSE"
 #  seg.e2 = "KIN HERE"
 #
-# class HL7::Message::Segment::Default < HL7::Message::Segment
-#  def initialize(raw_segment="", delims=[])
-#    segs = [] if (raw_segment == "")
-#    segs ||= raw_segment
-#    super( segs, delims )
-#  end
-#end
+class HL7::Message::Segment::Default < HL7::Message::Segment
+  def initialize(raw_segment="", delims=[])
+    segs = [] if (raw_segment == "")
+    segs ||= raw_segment
+    super( segs, delims )
+  end
+end
 
 # load our segments
-#Dir["#{File.dirname(__FILE__)}/segments/*.rb"].each { |ext| load ext }
+Dir["#{File.dirname(__FILE__)}/segments/*.rb"].each { |ext| load ext }
